Separate affinity mask from main process

+++ EnergyPerformance/Models/PersonaModel.cs
@@ -1,9 +1,7 @@
-using EnergyPerformance.Core.Contracts.Services;
 using EnergyPerformance.Helpers;
 using EnergyPerformance.Services;
 using System.Diagnostics;
 using System.Management;
-using System.Reflection.Metadata.Ecma335;
 
 namespace EnergyPerformance.Models;
 
@@ -227,11 +225,15 @@ public class PersonaModel
     /// Removes the Persona from the volatile and persistent memory
     /// </summary>
     /// <param name="personaName">The name of the Persona</param>
-    public async Task DeletePersona(string personaName)
+    public async Task<bool> DeletePersona(string personaName)
     {
+        if (IsEnabled && PersonaEnabled?.Path == personaName)
+            return false;
+
         _processMonitorService.RemoveWatcher(personaName);
         var res = _allPersonas.RemoveAll(persona => persona.Path.Equals(personaName, StringComparison.OrdinalIgnoreCase)) > 0;
         await _personaFileService.SaveFileAsync();
+        return true;
     }
 
     /// <summary>
